Make async-select loading-state look like in async-autocomplete
The reason for rendering the `MenuItem` components of the selected
values with `display: none` is they should not be visible in the
selects options, as the loading indicator is shown in that case.
The `MenuItem` components still need to be present in the DOM to be
visible within the input's closed state.